restructured python code
